move parse logic to query_io
